Moved admin routes; js_only.
