+++ assets/src/application/ui/display/espressoIcon/svgPaths.ts
@@ -5,6 +5,15 @@ import { SvgPath } from './types';
  * @link https://github.com/WordPress/gutenberg
  */
 export const svgPath: SvgPath = {
+	'admin-links':
+		'M17.74 2.76c1.68 1.69 1.68 4.41 0 6.1l-1.53 1.52c-1.12 1.12-2.7 1.47-4.14 ' +
+		'1.09l2.62-2.61.76-.77.76-.76c.84-.84.84-2.2 0-3.04-.84-.85-2.2-.85-3.04 ' +
+		'0l-.77.76-3.38 3.38c-.37-1.44-.02-3.02 1.1-4.14l1.52-1.53c1.69-1.68 4.42-1.68 ' +
+		'6.1 0zM8.59 13.43l5.34-5.34c.42-.42.42-1.1 0-1.52-.44-.43-1.13-.39-1.53 ' +
+		'0l-5.33 5.34c-.42.42-.42 1.1 0 1.52.44.43 1.13.39 1.52 0zm-.76 2.29l4.14-4.15c.38 ' +
+		'1.44.03 3.02-1.09 4.14l-1.52 1.53c-1.69 1.68-4.41 1.68-6.1 0-1.68-1.68-1.68-4.42 ' +
+		'0-6.1l1.53-1.52c1.12-1.12 2.7-1.47 4.14-1.1l-4.14 4.15c-.85.84-.85 2.2 0 ' +
+		'3.05.84.84 2.2.84 3.04 0z',
 	calculator:
 		'M 3 0 v20 h14 v-20 h-14 z ' +
 		'm 2.5 2.5 h9 v2 h-9 v-2 z ' +
@@ -47,6 +56,14 @@ export const svgPath: SvgPath = {
 		'c .54 -.12 .98 -.38 1.3 -.74 .3 -.36 .46 -.78 .46 -1.3 z ' +
 		'M 4 0 h2 v2 h-2 v-2 z ' +
 		'M 14 0 h2 v2 h-2 v-2 z',
+	edit:
+		'M13.89 3.39l2.71 2.72c.46.46.42 1.24.03 1.64l-8.01 ' +
+		'8.02-5.56 1.16 1.16-5.58s7.6-7.63 7.99-8.03c.39-.39 ' +
+		'1.22-.39 1.68.07zm-2.73 2.79l-5.59 5.61 1.11 1.11 ' +
+		'5.54-5.65zm-2.97 8.23l5.58-5.6-1.07-1.08-5.59 5.6z',
+	filter: 'M3 4.5v-2s3.34-1 7-1 7 1 7 1v2l-5 7.03v6.97s-1.22-.09-2.25-.59S8 16.5 8 16.5v-4.97z',
+	groups:
+		'M8.030 4.46c-0.29 1.28 0.55 3.46 1.97 3.46 1.41 0 2.25-2.18 1.96-3.46-0.22-0.98-1.080-1.63-1.96-1.63-0.89 0-1.74 0.65-1.97 1.63zM3.9 5.36c-0.25 1.080 0.47 2.93 1.67 2.93s1.92-1.85 1.67-2.93c-0.19-0.83-0.92-1.39-1.67-1.39s-1.48 0.56-1.67 1.39zM12.76 5.36c-0.25 1.080 0.47 2.93 1.66 2.93 1.2 0 1.92-1.85 1.67-2.93-0.19-0.83-0.92-1.39-1.67-1.39-0.74 0-1.47 0.56-1.66 1.39zM12.17 16.79l1.25-4.3c0.78-2.49-0.71-4.020-3.42-4.020-2.72 0-4.21 1.53-3.44 4.020l1.26 4.3c0.23 0.72 1.18 1.21 2.18 1.21 0.98 0 1.94-0.49 2.17-1.21zM6.070 9.16c-0.49 0.67-0.96 1.83-0.42 3.59l1.12 3.79c-0.34 0.2-0.77 0.31-1.2 0.31-0.85 0-1.65-0.41-1.85-1.030l-1.070-3.65c-0.65-2.11 0.61-3.4 2.92-3.4 0.27 0 0.54 0.020 0.79 0.060-0.1 0.1-0.2 0.22-0.29 0.33zM14.42 8.77c2.31 0 3.58 1.29 2.92 3.4l-1.070 3.65c-0.2 0.62-1 1.030-1.85 1.030-0.43 0-0.86-0.11-1.2-0.31l1.11-3.77c0.55-1.78 0.080-2.94-0.42-3.61-0.080-0.11-0.18-0.23-0.28-0.33 0.25-0.040 0.51-0.060 0.79-0.060z',
 	rem:
 		'M 10 1.02 ' +
 		'c -5.1 0 -8.75 4.04 -8.75 9 ' +
@@ -81,11 +98,7 @@ export const svgPath: SvgPath = {
 		'v -4.5 h -9 v 4.5 l 0.5 0.5 z M 11 6.5 v -3.5 h 2 ' +
 		'v 3 h -2 z M 4 11.5 v 0.5 h 12 v -0.5 z m 0 2 v 0.5 ' +
 		'h 12 v -0.5 z m 0 2 v 0.5 h 12 v -0.5 z',
-	edit:
-		'M13.89 3.39l2.71 2.72c.46.46.42 1.24.03 1.64l-8.01 ' +
-		'8.02-5.56 1.16 1.16-5.58s7.6-7.63 7.99-8.03c.39-.39 ' +
-		'1.22-.39 1.68.07zm-2.73 2.79l-5.59 5.61 1.11 1.11 ' +
-		'5.54-5.65zm-2.97 8.23l5.58-5.6-1.07-1.08-5.59 5.6z',
+
 	'ticket-alt':
 		'M20 6.38L18.99 9.2v-.01c-.52-.19-1.03-.16-' +
 		'1.53.08s-.85.62-1.04 1.14-.16 1.03.07 1.53c.24.5.62.84 1.15 ' +
@@ -105,11 +118,6 @@ export const svgPath: SvgPath = {
 		'0 1.25-.56 1.25-1.25v-1.75l.83.3c.12.43.36.78.71 1.04zM3.25 ' +
 		'17v-4c0-.41.34-.75.75-.75h.83l7.92 2.83V17c0 ' +
 		'.41-.34.75-.75.75H4c-.41 0-.75-.34-.75-.75z',
-	trash:
-		'M12 4h3c.6 0 1 .4 1 1v1H3V5c0-.6.5-1 1-1h3c.2-1.1 1.3-2 ' +
-		'2.5-2s2.3.9 2.5 2zM8 4h3c-.2-.6-.9-1-1.5-1S8.2 3.4 8 4zM4 ' +
-		'7h11l-.9 10.1c0 .5-.5.9-1 .9H5.9c-.5 0-.9-.4-1-.9L4 7z',
-	filter: 'M3 4.5v-2s3.34-1 7-1 7 1 7 1v2l-5 7.03v6.97s-1.22-.09-2.25-.59S8 16.5 8 16.5v-4.97z',
 	'list-view':
 		'M2 19h16c.55 0 1-.45 1-1V2c0-.55-.45-1-1-1H2c-.55 0-1 .45-1 1v16c0 ' +
 		'.55.45 1 1 1zM4 3c.55 0 1 .45 1 1s-.45 1-1 1-1-.45-1-1 .45-1 1-1zm13 ' +
@@ -117,15 +125,7 @@ export const svgPath: SvgPath = {
 		'0v2H6V7h11zM4 11c.55 0 1 .45 1 1s-.45 1-1 1-1-.45-1-1 .45-1 1-1zm13 ' +
 		'0v2H6v-2h11zM4 15c.55 0 1 .45 1 1s-.45 1-1 1-1-.45-1-1 .45-1 1-1zm13 ' +
 		'0v2H6v-2h11z',
-	'admin-links':
-		'M17.74 2.76c1.68 1.69 1.68 4.41 0 6.1l-1.53 1.52c-1.12 1.12-2.7 1.47-4.14 ' +
-		'1.09l2.62-2.61.76-.77.76-.76c.84-.84.84-2.2 0-3.04-.84-.85-2.2-.85-3.04 ' +
-		'0l-.77.76-3.38 3.38c-.37-1.44-.02-3.02 1.1-4.14l1.52-1.53c1.69-1.68 4.42-1.68 ' +
-		'6.1 0zM8.59 13.43l5.34-5.34c.42-.42.42-1.1 0-1.52-.44-.43-1.13-.39-1.53 ' +
-		'0l-5.33 5.34c-.42.42-.42 1.1 0 1.52.44.43 1.13.39 1.52 0zm-.76 2.29l4.14-4.15c.38 ' +
-		'1.44.03 3.02-1.09 4.14l-1.52 1.53c-1.69 1.68-4.41 1.68-6.1 0-1.68-1.68-1.68-4.42 ' +
-		'0-6.1l1.53-1.52c1.12-1.12 2.7-1.47 4.14-1.1l-4.14 4.15c-.85.84-.85 2.2 0 ' +
-		'3.05.84.84 2.2.84 3.04 0z',
+
 	'editor-unlink':
 		'M17.74 2.26c1.68 1.69 1.68 4.41 0 6.1l-1.53 1.52c-.32.33-.69.58-1.08.77L13 ' +
 		'10l1.69-1.64.76-.77.76-.76c.84-.84.84-2.2 0-3.04-.84-.85-2.2-.85-3.04 ' +
@@ -134,4 +134,8 @@ export const svgPath: SvgPath = {
 		'13l.74 2.35-1.38 1.39c-1.69 1.68-4.41 1.68-6.1 0-1.68-1.68-1.68-4.42 ' +
 		'0-6.1l1.39-1.38L7 10l-.69.64-1.52 1.53c-.85.84-.85 2.2 0 3.04.84.85 2.2.85 ' +
 		'3.04 0zM18 16l-8-6 6 8zm-4 2l-4-8 2 8h2zm4-4l-8-4 8 2v2z',
+	trash:
+		'M12 4h3c.6 0 1 .4 1 1v1H3V5c0-.6.5-1 1-1h3c.2-1.1 1.3-2 ' +
+		'2.5-2s2.3.9 2.5 2zM8 4h3c-.2-.6-.9-1-1.5-1S8.2 3.4 8 4zM4 ' +
+		'7h11l-.9 10.1c0 .5-.5.9-1 .9H5.9c-.5 0-.9-.4-1-.9L4 7z',
 };
