+++ assets/src/domain/eventEditor/ui/datetimes/datesList/cardView/DateDetails.tsx
@@ -1,4 +1,4 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 import { __ } from '@wordpress/i18n';
 
 import InfinitySymbol from '@application/valueObjects/InfinitySymbol';
@@ -14,34 +14,33 @@ interface DateDetailsProps {
 }
 
 const DateDetails: React.FC<DateDetailsProps> = ({ datetime, updateDatetime }) => {
-	return useMemo(() => {
-		const details = [
-			{
-				id: 'ee-event-date-sold',
-				label: __('sold'),
-				value: datetime.sold || 0,
+	const details = [
+		{
+			id: 'ee-event-date-sold',
+			label: __('sold'),
+			value: datetime.sold || 0,
+		},
+		{
+			id: 'ee-event-date-capacity',
+			label: __('capacity'),
+			value: <InfinitySymbol value={datetime.capacity} asInt />,
+			onChange: (cap: any) => {
+				const capacity = parseInfinity(cap, true);
+				updateDatetime({ capacity });
+				/**
+				 * @TODO update related tickets capacity
+				 */
 			},
-			{
-				id: 'ee-event-date-capacity',
-				label: __('capacity'),
-				value: <InfinitySymbol value={datetime.capacity} asInt />,
-				onChange: (cap: any) => {
-					const capacity = parseInfinity(cap, true);
-					updateDatetime({ capacity });
-					/**
-					 * @TODO update related tickets capacity
-					 */
-				},
-			},
-			{
-				id: 'ee-event-date-registrations',
-				className: 'ee-has-tooltip',
-				label: __('reg list'),
-				value: <DateRegistrationsLink datetime={datetime} />,
-			},
-		];
-		return <EntityDetailsPanel details={details} className='ee-editor-date-details-sold-rsrvd-cap-div' />;
-	}, [datetime.capacity, datetime.sold]);
+		},
+		{
+			id: 'ee-event-date-registrations',
+			className: 'ee-has-tooltip',
+			label: __('reg list'),
+			value: <DateRegistrationsLink datetime={datetime} />,
+		},
+	];
+
+	return <EntityDetailsPanel details={details} className='ee-editor-date-details-sold-rsrvd-cap-div' />;
 };
 
 export default DateDetails;
